Move dracut code from cliwrap to initramfs module
Prep for handling kernel-install in a different way, where
we add a drop-in that is invoked by the existing binary instead
of intercepting it entirely.

No functional changes intended (yet).
